backend_auth/requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
celery==5.5.3
redis==6.4.0
djangorestframework==3.16.1
aiogram~=3.22.0
requests~=2.32.5
pillow~=11.3.0
beautifulsoup4~=4.13.4
drf_camel_case==1.0.2
djangorestframework-simplejwt==5.5.1
cryptography==45.0.7
django-cors-headers==4.9.0
whitenoise==6.11.0
